Creating tusayan_whiteware.py to ease loading of dataset
